* 'master' of github.com:AY2425S1-CS2113-W13-4/tp:
  Modify text-ui-test
  Fix test cases
  Update input.txt
  Update .gitignore
  Add support for list_grade command
  Modify GradeList.deleteGrade
  Fix ComponentList.toString() formatting
  Add input validation to Commons
  Validate matricNumber as per regex
  Remove redundant .toUpperCase() call
  Change weightage from double to int
  Add setup section to DG
  Regenerate .png
  Remove Hashmap class
  Remove Hashmap class
  Modify DG to align format
  Add Setup.puml
  Remove Scanner object
  Abstract setup ref frame
  Remove HashMap object

# Conflicts:
#	src/main/java/tutorlink/storage/ComponentStorage.java

+++ src/main/java/tutorlink/component/Component.java
@@ -4,9 +4,9 @@ package tutorlink.component;
 public class Component {
     private String name;
     private double maxScore;
-    private double weight;
+    private int weight;
 
-    public Component(String name, double maxScore, double weight) {
+    public Component(String name, double maxScore, int weight) {
         this.name = name;
         this.maxScore = maxScore;
         this.weight = weight;
